not number char bug correction

+++ push_swap/case_3.c
@@ -6,7 +6,7 @@
 /*   By: umartin- <umartin-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/10 12:47:52 by umartin-          #+#    #+#             */
-/*   Updated: 2022/03/10 14:35:43 by umartin-         ###   ########.fr       */
+/*   Updated: 2022/03/14 17:26:09 by umartin-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,25 +19,25 @@ void	case_3(t_list	**sta)
 	temp = *sta;
 	if (temp == NULL || temp->next == NULL || temp->next->next == NULL)
 		return ;
-	else if ((temp->content > temp->next->content)
-		&& (temp->next->content > temp->next->next->content)
-		&& (temp->content > temp->next->next->content))
+	else if (((int)temp->content > (int)temp->next->content)
+		&& ((int)temp->next->content > (int)temp->next->next->content)
+		&& ((int)temp->content > (int)temp->next->next->content))
 		case_3_1(sta);
-	else if ((temp->content > temp->next->content)
-		&& (temp->next->content < temp->next->next->content)
-		&& (temp->content < temp->next->next->content))
+	else if (((int)temp->content > (int)temp->next->content)
+		&& ((int)temp->next->content < (int)temp->next->next->content)
+		&& ((int)temp->content < (int)temp->next->next->content))
 		psw_sa(sta);
-	else if ((temp->content > temp->next->content)
-		&& (temp->next->content < temp->next->next->content)
-		&& (temp->content > temp->next->next->content))
+	else if (((int)temp->content > (int)temp->next->content)
+		&& ((int)temp->next->content < (int)temp->next->next->content)
+		&& ((int)temp->content > (int)temp->next->next->content))
 		psw_ra(sta);
-	else if ((temp->content < temp->next->content)
-		&& (temp->next->content > temp->next->next->content)
-		&& (temp->content > temp->next->next->content))
+	else if (((int)temp->content < (int)temp->next->content)
+		&& ((int)temp->next->content > (int)temp->next->next->content)
+		&& ((int)temp->content > (int)temp->next->next->content))
 		psw_rra(sta);
-	else if ((temp->content < temp->next->content)
-		&& (temp->next->content > temp->next->next->content)
-		&& (temp->content < temp->next->next->content))
+	else if (((int)temp->content < (int)temp->next->content)
+		&& ((int)temp->next->content > (int)temp->next->next->content)
+		&& ((int)temp->content < (int)temp->next->next->content))
 		case_3_2(sta);
 }
